Warn that the error reading default POPT config file may be bogus.
This is the result of a bug in the POPT library [1] which occurs
when the /etc/popt.d directory is present but empty.  POPT returns
a bogus error indication in this case which cannot be definitively
detected (because it may be legitimate or incorrectly returned with
a random value of 'errno').

Refs: [1] http://rpm5.org/cvs/tktview?tn=88

git-svn-id: svn://daper.net/moc/trunk@2764 910807d9-36e0-0310-a014-e9ea483e2ba4